pyproject.toml
[project]
name = "website"
version = "0.5.6"
description = ""
authors = [
    {name = "kettleofketchup@gmail.com",email = "kettleofketchup@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"

package-mode = false
dependencies = [

    "django (>=5.2,<6.0)",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "python-social-auth[django]>=0.3.6",
    "requests>=2.32.3",
    "social-auth-app-django>=5.4.3",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-jinja (>=2.11.0,<3.0.0)",
    "invoke (>=2.2.0,<3.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "black (>=25.1.0,<26.0.0)",
    "alive-progress (>=3.2.0,<4.0.0)",
    "discord (>=2.3.2,<3.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "faker (>=37.3.0,<38.0.0)",
    "certbot-nginx (>=4.1.1,<5.0.0)",
    "semver (>=3.0.4,<4.0.0)",
    "poetry (>=2.2.0,<3.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "django-cacheops (>=7.2,<8.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.7.0,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "celery[redis] (>=5.3,<6.0)",
    "channels (>=4.0.0,<5.0.0)",
    "channels-redis (>=4.2.0,<5.0.0)",
    "daphne (>=4.0.0,<5.0.0)",
    "pynacl (>=1.5.0,<2.0.0)",
    "nh3 (>=0.3.2,<0.4.0)",
]



[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"
pytest = "^8.4.2"
mkdocs-material = "^9.7.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
src_paths = ["backend", "."]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
